functionality
add functionality to the design
