+++ src/App.vue
@@ -1,15 +1,17 @@
 <script setup>
 import { ref, onMounted, onUnmounted, watch } from 'vue'
-import router from './router';
 import { useRoute } from 'vue-router'
 
 
 const route = useRoute()
 const scrolled = ref(false)
 const inHome = ref(true)
+const svgPrimary = ref(false)
 
 const handleScroll = () => {
   scrolled.value = window.scrollY > 50
+  svgPrimary.value = window.scrollY > 50
+
 }
 
 
@@ -17,7 +19,7 @@ const handleScroll = () => {
 onMounted(() => {
   window.addEventListener('scroll', handleScroll)
   inHome.value = true
-  console.log(inHome.value)
+
 })
 
 onUnmounted(() => {
@@ -39,18 +41,28 @@ watch(route, (newRoute) => {
 </script>
 
 <template>
-    <div v-if="inHome" :class="[' flex gap-4 justify-between px-80 py-5 text-xl fixed m-auto w-full top-0 left-0 transition-colors duration-300', { 'bg-white shadow-md text-primary': scrolled, 'bg-transparent text-white': !scrolled }]"
+    <div v-if="inHome" :class="[' py-5 text-xl fixed w-full top-0 transition-colors duration-300',
+     { 'bg-white shadow-md text-primary': scrolled,
+      'bg-transparent text-white': !scrolled }]"
       >
-        <h1 class="text-4xl font-Karla font-extrabold">Meyawo</h1>
-        <div class="flex gap-4 justify-center items-center font-Karla font-thin">
-          <RouterLink to="/">Home</RouterLink>
-          <p>About</p>
-          <p>Portfolio</p>
-          <p>Testmonial</p>
-          <p>Blog</p>
-          <p>Contact</p>
-          <RouterLink to="/about" class="p-2 px-4 bg-primary text-white rounded-sm" >Components</RouterLink>
-          <svg xmlns="http://www.w3.org/2000/svg" height="32" width="24" viewBox="0 0 384 512"><path fill="#ffffff"  opacity="0.8" d="M342.6 150.6c12.5-12.5 12.5-32.8 0-45.3s-32.8-12.5-45.3 0L192 210.7 86.6 105.4c-12.5-12.5-32.8-12.5-45.3 0s-12.5 32.8 0 45.3L146.7 256 41.4 361.4c-12.5 12.5-12.5 32.8 0 45.3s32.8 12.5 45.3 0L192 301.3 297.4 406.6c12.5 12.5 32.8 12.5 45.3 0s12.5-32.8 0-45.3L237.3 256 342.6 150.6z"/></svg>
+        <div class="flex justify-between gap-8 w-3/4 mx-auto">
+          <a href="#">
+            <h1 class="text-4xl font-Karla font-extrabold cursor-pointer">Meyawo</h1>
+          </a>
+          <div class="hidden xl:flex gap-8 justify-center items-center font-Karla font-thin">
+          
+            <ul class="flex space-x-8 justify-center ">
+              <li><a href="#home" class=" cursor-pointer hover:text-hoverd">Home</a></li>
+              <li><a href="#about" class=" cursor-pointer hover:text-hoverd">About</a></li>
+              <li><a href="#portfolio" class=" cursor-pointer hover:text-hoverd">Portfolio</a></li>
+              <li><a href="#testomanial" class=" cursor-pointer hover:text-hoverd">Testmonial</a></li>
+              <li><a href="#blog" class=" cursor-pointer hover:text-hoverd">Blog</a></li>
+              <li><a href="#contact" class=" cursor-pointer hover:text-hoverd">Contact</a></li>
+            </ul>
+            <RouterLink to="/components" class="p-2 px-4 bg-primary text-white rounded-sm hover:bg-hoverd" >Components</RouterLink>
+          </div>
+          <svg v-if="!svgPrimary" class=" xl:hidden fill-white hover:fill-primary cursor-pointer" xmlns="http://www.w3.org/2000/svg" height="32" width="24" viewBox="0 0 384 512"><path d="M0 96C0 78.3 14.3 64 32 64l384 0c17.7 0 32 14.3 32 32s-14.3 32-32 32L32 128C14.3 128 0 113.7 0 96zM0 256c0-17.7 14.3-32 32-32l384 0c17.7 0 32 14.3 32 32s-14.3 32-32 32L32 288c-17.7 0-32-14.3-32-32zM448 416c0 17.7-14.3 32-32 32L32 448c-17.7 0-32-14.3-32-32s14.3-32 32-32l384 0c17.7 0 32 14.3 32 32z"/></svg>
+          <svg v-else class="  xl:hidden fill-primary hover:fill-hoverd cursor-pointer" xmlns="http://www.w3.org/2000/svg" height="32" width="24" viewBox="0 0 384 512"><path d="M0 96C0 78.3 14.3 64 32 64l384 0c17.7 0 32 14.3 32 32s-14.3 32-32 32L32 128C14.3 128 0 113.7 0 96zM0 256c0-17.7 14.3-32 32-32l384 0c17.7 0 32 14.3 32 32s-14.3 32-32 32L32 288c-17.7 0-32-14.3-32-32zM448 416c0 17.7-14.3 32-32 32L32 448c-17.7 0-32-14.3-32-32s14.3-32 32-32l384 0c17.7 0 32 14.3 32 32z"/></svg>
         </div>
     </div>
